Refactor search paths collection (#188)
Before this refactor we generally tried to calculate search paths, to ensure that we correctly identified the "external/" and "bazel-out/" directories without having to parse strings. This has the benefit of technically allowing the "external/" and "bazel-out/" directories to exist in the workspace.

There are some issues with this approach though. The main one being that custom rules that set search paths aren't accounted for properly. If we instead leverage the information in `CcInfo.compilation_context` we support all sorts of rules out of the box... but have to resort parsing strings.

We were already parsing the strings for the `local_includes`, and will need to for framework includes in the future. I believe the more generic code that should just work is worth giving up the ability to have "external/" and "bazel-out/" directories in the workspace.