Refactor input collection (#199)
We no longer use an additional aspect to collect input files. This is needed in order to properly classify `extra_files`. Before this change, while trying to implement resource bundle targets, the resources would always be classified as extra files, because of the transitive nature of  how `input_files_apsect` collected the `other` field. Now we are able to leverage the information in `XcodeProjInfo` to determine the correct classification of all input files.

This also now includes intermediate generated files. I personally think this is a regression (though others might disagree), but I don't know of an easy way to exclude them at this time (if we even want to).

Finally, `generated_inputs` and `extra_files` have been moved into the `input_files` `struct`.