Fix over-collection of Xcode targets (#212)
Before this change we would create Xcode targets and set dependencies to them along every rule attribute. This was partially a side effect of the changes in input collection that requires us to propagate the aspect along all attributes.

Now we only include transitive information for attributes that should create Xcode targets.