List sources in the same order that Bazel sees them (#265)
This helps ensure that the files are compiled in the same order that bazel compiles them in. I've encountered issues in the past for Swift where the order matters. This will also help Build with Bazel to have Xcode try to "compile" things in the same order that Bazel actually is.