Create generated file parent directories during Setup (#297)
If the parent directories of generated files don't exist before an Xcode project is opened, then after the files are created Xcode still won't see them until the project is closed and opened again.

As a step towards making that experience better we will now pre-create those directories in the Setup target. In a later change we will invoke this target during project generation to put the project in a better state from the start.