Merge `createFetchExternalReposScript` and `createGenerateFilesScript` (#356)
The differences between these functions were minor, and Build with Bazel is easier to implement if they are one function. While doing this I decided to let `Fetch External Repositories` have the same `bazel build` invocation, since `--output_group=generated_inputs` is a no-op (and actually does _less_ than the version without it), and it again makes the Build with Bazel change smaller.